stats4phc: Performance evaluation for the prognostic value of predictive models intended to support personalized healthcare through predictiveness curves and positive / negative predictive values

vignette("stats4phc.Rmd")

See also

Author

Maintainer: Ondrej Slama ondrej.slama@roche.com

Authors:

Darrick Shen shend9@gene.com

Verena Steffen steffen.verena@gene.com

Doug Kelkhoff doug.kelkhoff@roche.com

Michel Friesenhahn friesem1@gene.com

Christina Rabe rabec@gene.com

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]